Add date e Administrador Comum

+++ src/main/java/ufrn/br/web/repositoreis/PessoaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import ufrn.br.web.model.Pessoa;
 
 @Repository
-public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
+public interface PessoaRepository extends JpaRepository<Pessoa, Integer> {
 }
